myproject/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
channels==3.0.0
channels_redis==4.2.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
command-not-found==0.3
constantly==23.10.4
cryptography==43.0.3
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.1.12
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djongo==1.3.7
dnspython==2.7.0
filelock==3.6.0
gunicorn==23.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
mongoengine==0.29.1
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.10.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.0
reportlab==3.6.8
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
service-identity==24.2.0
six==1.16.0
sqlparse==0.2.4
systemd-python==234
tomli==2.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
urllib3==1.26.5
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==7.1.1
